Star database
A list of celestial objects, alongside their celestial coordinates, apparent magnitude, and parallax (useful to calculate distance). This file was queried from the Simbad database, making sure their apparent magnitude was lower than 6.5, making them visible to the naked eye in a clear night, and that the parallax was not null.